+++ Amethyst/Worlds/Section.cs
@@ -8,7 +8,7 @@ internal sealed class Section
     private readonly byte[] blocksLight = new byte[2048];
     private readonly byte[] skyLight = new byte[2048];
 
-    public Block GetBlock(long x, long y, long z)
+    public Block GetBlock(int x, int y, int z)
     {
         var index = AsIndex(x, y, z) * 2;
 
@@ -18,7 +18,7 @@ internal sealed class Section
         return new Block(type, metadata);
     }
 
-    public void SetBlock(Block block, long x, long y, long z)
+    public void SetBlock(Block block, int x, int y, int z)
     {
         var index = AsIndex(x, y, z) * 2;
         var type = block.Type;
@@ -27,13 +27,13 @@ internal sealed class Section
         blocks[index + 1] = (byte) (type >> 4);
     }
 
-    public byte GetSkyLight(long x, long y, long z)
+    public byte GetSkyLight(int x, int y, int z)
     {
         var index = AsIndex(x, y, z);
         return skyLight[index];
     }
 
-    public void SetSkyLight(byte value, long x, long y, long z)
+    public void SetSkyLight(byte value, int x, int y, int z)
     {
         var index = AsIndex(x, y, z);
         skyLight[index] = value;
@@ -50,7 +50,7 @@ internal sealed class Section
         return (blocks, blocksLight, skyLight);
     }
 
-    private static long AsIndex(long x, long y, long z)
+    private static int AsIndex(int x, int y, int z)
     {
         return (y & 0xF) << 8 | (z & 0xF) << 4 | x & 0xF;
     }
